AT-2649 Scan band = int

+++ tests/integration/tmc/test_end_to_end.py
@@ -295,7 +295,7 @@ def _(telescope_handlers, receptor_ids, pb_and_eb_ids):
 
 @when(
     parsers.cfparse(
-        "configure it for a band {scan_band:Number} scan", extra_types={"Number": float}
+        "configure it for a band {scan_band:Number} scan", extra_types={"Number": int}
     )
 )
 def _(telescope_handlers, receptor_ids, scan_band):
